+++ docx-core/src/documents/elements/table_of_contents.rs
@@ -1,8 +1,43 @@
+use serde::ser::{SerializeStruct, Serializer};
 use serde::Serialize;
 
 use crate::documents::*;
 use crate::types::*;
 use crate::xml_builder::*;
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum TocContent {
+    Paragraph(Box<Paragraph>),
+    Table(Box<Table>),
+}
+
+impl Serialize for TocContent {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        match *self {
+            TocContent::Paragraph(ref p) => {
+                let mut t = serializer.serialize_struct("Paragraph", 2)?;
+                t.serialize_field("type", "paragraph")?;
+                t.serialize_field("data", p)?;
+                t.end()
+            }
+            TocContent::Table(ref c) => {
+                let mut t = serializer.serialize_struct("Table", 2)?;
+                t.serialize_field("type", "table")?;
+                t.serialize_field("data", c)?;
+                t.end()
+            }
+        }
+    }
+}
+
+#[derive(Serialize, Debug, Clone, PartialEq, Default)]
+pub struct TableOfContentsReviewData {
+    pub author: String,
+    pub date: String,
+}
 
 // https://c-rex.net/projects/samples/ooxml/e1/Part4/OOXML_P4_DOCX_TOCTOC_topic_ID0ELZO1.html
 // This struct is only used by writers
@@ -10,15 +45,32 @@ use crate::xml_builder::*;
 pub struct TableOfContents {
     pub instr: InstrToC,
     pub items: Vec<TableOfContentsItem>,
+    // don't use
     pub auto: bool,
     pub dirty: bool,
     pub alias: Option<String>,
     pub page_ref_placeholder: Option<String>,
+    // it is inserted in before toc.
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    pub before_contents: Vec<TocContent>,
+    // it is inserted in after toc.
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    pub after_contents: Vec<TocContent>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub delete: Option<TableOfContentsReviewData>,
 }
 
 impl TableOfContents {
     pub fn new() -> Self {
         Self::default()
+    }
+
+    pub fn with_instr_text(s: &str) -> Self {
+        let instr = InstrToC::with_instr_text(s);
+        Self {
+            instr,
+            ..Self::default()
+        }
     }
 
     pub fn heading_styles_range(mut self, start: usize, end: usize) -> Self {
@@ -41,6 +93,14 @@ impl TableOfContents {
         self
     }
 
+    pub fn delete(mut self, author: impl Into<String>, date: impl Into<String>) -> Self {
+        self.delete = Some(TableOfContentsReviewData {
+            author: author.into(),
+            date: date.into(),
+        });
+        self
+    }
+
     // pub fn tc_field_level_range(mut self, start: usize, end: usize) -> Self {
     //     self.instr = self.instr.tc_field_level_range(start, end);
     //     self
@@ -60,6 +120,27 @@ impl TableOfContents {
         self.dirty = true;
         self
     }
+
+    pub fn add_before_paragraph(mut self, p: Paragraph) -> Self {
+        self.before_contents
+            .push(TocContent::Paragraph(Box::new(p)));
+        self
+    }
+
+    pub fn add_after_paragraph(mut self, p: Paragraph) -> Self {
+        self.after_contents.push(TocContent::Paragraph(Box::new(p)));
+        self
+    }
+
+    pub fn add_before_table(mut self, t: Table) -> Self {
+        self.before_contents.push(TocContent::Table(Box::new(t)));
+        self
+    }
+
+    pub fn add_after_table(mut self, t: Table) -> Self {
+        self.after_contents.push(TocContent::Table(Box::new(t)));
+        self
+    }
 }
 
 impl BuildXML for TableOfContents {
@@ -69,23 +150,70 @@ impl BuildXML for TableOfContents {
             p = p.alias(alias);
         }
         if self.items.is_empty() {
-            let p1 = Paragraph::new().add_run(
-                Run::new()
-                    .add_field_char(FieldCharType::Begin, true)
-                    .add_instr_text(InstrText::TOC(self.instr.clone()))
-                    .add_field_char(FieldCharType::Separate, false),
-            );
-            let p2 = Paragraph::new().add_run(Run::new().add_field_char(FieldCharType::End, false));
-
-            XMLBuilder::new()
+            let mut b = XMLBuilder::new()
                 .open_structured_tag()
                 .add_child(&p)
-                .open_structured_tag_content()
-                .add_child(&p1)
-                .add_child(&p2)
-                .close()
-                .close()
-                .build()
+                .open_structured_tag_content();
+
+            for c in self.before_contents.iter() {
+                match c {
+                    TocContent::Paragraph(p) => {
+                        b = b.add_child(p);
+                    }
+                    TocContent::Table(t) => {
+                        b = b.add_child(t);
+                    }
+                }
+            }
+
+            let p1 = if let Some(ref del) = self.delete {
+                Paragraph::new().add_delete(
+                    Delete::new().author(&del.author).date(&del.date).add_run(
+                        Run::new()
+                            .add_field_char(FieldCharType::Begin, true)
+                            .add_delete_instr_text(DeleteInstrText::TOC(self.instr.clone()))
+                            .add_field_char(FieldCharType::Separate, false),
+                    ),
+                )
+            } else {
+                Paragraph::new().add_run(
+                    Run::new()
+                        .add_field_char(FieldCharType::Begin, true)
+                        .add_instr_text(InstrText::TOC(self.instr.clone()))
+                        .add_field_char(FieldCharType::Separate, false),
+                )
+            };
+            b = b.add_child(&p1);
+
+            let p2 = Paragraph::new().add_run(Run::new().add_field_char(FieldCharType::End, false));
+            if self.after_contents.is_empty() {
+                b = b.add_child(&p2);
+            } else {
+                for (i, c) in self.after_contents.iter().enumerate() {
+                    match c {
+                        TocContent::Paragraph(p) => {
+                            // Merge paragraph
+                            if i == 0 {
+                                let mut new_p = p.clone();
+                                new_p.children.insert(
+                                    0,
+                                    ParagraphChild::Run(Box::new(
+                                        Run::new().add_field_char(FieldCharType::End, false),
+                                    )),
+                                );
+                                b = b.add_child(&new_p)
+                            } else {
+                                b = b.add_child(p);
+                            }
+                        }
+                        TocContent::Table(t) => {
+                            b = b.add_child(t);
+                        }
+                    }
+                }
+            }
+
+            b.close().close().build()
         } else {
             let items: Vec<TableOfContentsItem> = self
                 .items
@@ -100,14 +228,37 @@ impl BuildXML for TableOfContents {
                     item
                 })
                 .collect();
-            XMLBuilder::new()
+
+            let mut b = XMLBuilder::new()
                 .open_structured_tag()
                 .add_child(&p)
-                .open_structured_tag_content()
-                .add_child(&items)
-                .close()
-                .close()
-                .build()
+                .open_structured_tag_content();
+
+            for c in self.before_contents.iter() {
+                match c {
+                    TocContent::Paragraph(p) => {
+                        b = b.add_child(p);
+                    }
+                    TocContent::Table(t) => {
+                        b = b.add_child(t);
+                    }
+                }
+            }
+
+            b = b.add_child(&items);
+
+            for c in self.after_contents.iter() {
+                match c {
+                    TocContent::Paragraph(p) => {
+                        b = b.add_child(p);
+                    }
+                    TocContent::Table(t) => {
+                        b = b.add_child(t);
+                    }
+                }
+            }
+
+            b.close().close().build()
         }
     }
 }