+++ docx-core/src/errors/mod.rs
@@ -2,6 +2,8 @@ use thiserror::Error;
 
 #[derive(Error, Debug)]
 pub enum DocxError {
+    #[error("FromStr error.{0}")]
+    ConvertError(String),
     #[error("Failed to write XML to buffer.")]
     EmitterError(#[from] xml::writer::Error),
     #[error("Failed to zip XML documents.")]