+++ docx-wasm/js/paragraph.ts
@@ -151,6 +151,11 @@ export class Paragraph {
     return this;
   }
 
+  outlineLevel(v: number) {
+    this.property = { ...this.property, outlineLvl: v };
+    return this;
+  }
+
   paragraphPropertyChange(propertyChange: ParagraphPropertyChange) {
     this.property.paragraphPropertyChange = propertyChange;
     return this;