+++ docx-wasm/js/settings.ts
@@ -7,6 +7,7 @@ export class Settings {
   _docId: string | null = null;
   _docVars: DocVar[] = [];
   _defaultTabStop = 840;
+  _adjustLineHeightInTable = false;
 
   docId(id: string) {
     this._docId = id;
@@ -20,6 +21,11 @@ export class Settings {
 
   addDocVar(name: string, val: string) {
     this._docVars.push({ name, val });
+    return this;
+  }
+
+  adjustLineHeightInTable() {
+    this._adjustLineHeightInTable = true;
     return this;
   }
 }